requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
pytest==8.3.4
pytest-asyncio==0.23.5
python-multipart==0.0.9
notion-client==2.2.1
google-cloud-logging==3.8.0
